Refactor jsonwebtoken adapter into srp modules
